login v1

+++ backend/rest-prototype/src/main/java/com/manos/prototype/service/UserServiceImpl.java
@@ -14,6 +14,7 @@ import com.manos.prototype.dao.UserDao;
 import com.manos.prototype.dto.UserDTO;
 import com.manos.prototype.entity.Role;
 import com.manos.prototype.entity.User;
+import com.manos.prototype.exception.EntityNotFoundException;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -31,19 +32,31 @@ public class UserServiceImpl implements UserService {
 	@Autowired
 	private RoleDao roleDao;
 	
-//	@Autowired
-//	private BCryptPasswordEncoder passwordEncoder;
+	@Autowired
+	private BCryptPasswordEncoder passwordEncoder;
 
 	@Override
 	@Transactional
 	public User findByUserName(String userName) {
 		// check the database if the user already exists
-		return userDao.findByUserName(userName);
+		User user = userDao.findByUserName(userName);
+		return user;
 	}
 
 	@Override
 	@Transactional
-	public void saveUser(User user, UserDTO userDTO) {
+	public User saveUser(UserDTO userDTO) {
+		if (findByUserName(userDTO.getUserName()) != null) {
+			throw new EntityNotFoundException("User name already exists - " + userDTO.getUserName(), new Exception());
+		}
+		User tempUser = new User();
+		tempUser.setId(Long.parseLong("0"));
+		tempUser.setEmail(userDTO.getEmail());
+		tempUser.setFirstName(userDTO.getFirstName());
+		tempUser.setLastName(userDTO.getLastName());
+		tempUser.setPassword(passwordEncoder.encode(userDTO.getPassword()));
+		tempUser.setUserName(userDTO.getUserName());
+		
 		// give roles to the user
 		Collection<Role> tempColl = new ArrayList<>(); 
 		tempColl.add( roleDao.findRoleByName(userDTO.getRole().toString()) );
@@ -51,14 +64,20 @@ public class UserServiceImpl implements UserService {
 			tempColl.add(roleDao.findRoleByName("ROLE_EMPLOYEE")); // give default role
 		}
 		
-		user.setRoles(tempColl);
+		tempUser.setRoles(tempColl);
 		 // save user in the database
-		userDao.saveUser(user);
+		try {
+			userDao.saveUser(tempUser);			
+		} catch (Exception e) {
+			throw new EntityNotFoundException(e.getCause().getLocalizedMessage());
+		}
+		return tempUser;
 	}
 
 	@Override
 	@Transactional
 	public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
+		System.out.println("test "+userName);
 		User user = userDao.findByUserName(userName);
 		if (user == null) {
 			throw new UsernameNotFoundException("Invalid username or password.");
@@ -81,19 +100,27 @@ public class UserServiceImpl implements UserService {
 
 	@Override
 	@Transactional
-	public List<User> getUsers() {
-		return userDao.getUsers();
-	}
-
-	@Override
-	@Transactional
 	public User getUser(long userId) {
 		return userDao.getUser(userId);
 	}
 
 	@Override
 	@Transactional
-	public void updateUser(User user, UserDTO userDTO) {
+	public User updateUser(UserDTO userDTO) {
+		
+		User tempUser = getUser(userDTO.getId());
+		
+		if (tempUser == null) {
+			throw new EntityNotFoundException("User id not found - " + userDTO.getId());
+		}
+//		User user = new User();
+		tempUser.setId((long)userDTO.getId());
+		tempUser.setEmail(userDTO.getEmail());
+		tempUser.setFirstName(userDTO.getFirstName());
+		tempUser.setLastName(userDTO.getLastName());
+		tempUser.setPassword(passwordEncoder.encode(userDTO.getPassword()));
+		tempUser.setUserName(userDTO.getUserName());
+		
 		// give roles to the user
 		Collection<Role> tempColl = new ArrayList<>(); 
 		tempColl.add( roleDao.findRoleByName(userDTO.getRole().toString()) );
@@ -101,14 +128,23 @@ public class UserServiceImpl implements UserService {
 			tempColl.add(roleDao.findRoleByName("ROLE_EMPLOYEE")); // give default role
 		}
 		
-		user.setRoles(tempColl);
+		tempUser.setRoles(tempColl);
 		 // save user in the database
-		userDao.updateUser(user);
+		try {
+			userDao.updateUser(tempUser);
+		} catch (Exception e) {
+			throw new EntityNotFoundException(e.getCause().getLocalizedMessage());
+		}
+		return tempUser;
 	}
 
 	@Override
 	@Transactional
 	public void deleteUser(long userId) {
+		User tempUser = getUser(userId);
+		if (tempUser == null) {
+			throw new EntityNotFoundException("User id not found - " + userId);
+		}
 		userDao.deleteUser(userId);
 	}
 }
